Fix always-failing `Invoice::upcoming(...)`
`Invoice::upcoming(...)` would always fail. The API call returns an
invoice that doesn't actually exist, and because it doesn't exist, it
doesn't have an `"id"` field in the response.

> Note that when you are viewing an upcoming invoice, you are simply
> viewing a preview – the invoice has not yet been created. As such, the
> upcoming invoice will not show up in invoice listing calls, and you
> cannot use the API to pay or edit the invoice. If you want to change the
> amount that your customer will be billed, you can add, remove, or update
> pending invoice items, or update the customer’s discount.

The lack of an `"id"` field causes the parsing of an `Invoice` to fail.

We can't do `struct Invoice { id: Option<InvoiceId> }` because the
`Object` trait expects the ID of the object to not be an option.

It seems like the least terrible option is to set `#[serde(default)]` on
the ID field and provide an empty string as an ID. While this isn't
great, it seems like the best of the bad options.