fix: Re-export checkout_session_ext
checkout_session_ext.rs contains the RetrieveCheckoutSessionLineItems
struct. Without this re-export it is impossible to create an instance
and thus call CheckoutSession::retrieve_line_items.
There are other *_ext.rs files such as balance_ext.rs that are not
re-exported but they only contain trait implementations. I'm asssuming
the RetrieveCheckoutSessionLineItems struct got added to the
checkout_session_ext.rs file when it also contained just a trait
implementation but wasn't re-exported to reflect this addition.
Could be worth it to re-export everything on principle to avoid future
mistakes.