fix: correct docs host and path for API references
The Stripe documentation now is hosted on https://docs.stripe.com/api. The old https://stripe.com/docs/api still works but gets redirected to the new one.
The issue is with how we use to generate documentation link on url_finder.rs. On that file, we look for the load the docs and find the window.__INITIAL_STATE__, which contains relative links to the resources.
We are using those relative links as if they were relative from https://stripe.com, but now they are relative from https://docs.stripe.com  which causes the links to not be correctly built.
This is the issue that is causing the verify-codegen CI job to fail.
By updating it to the right path, it should stop generating incorrect path to the docs.