chore: align and pin Rust versions
Currently, new Rust releases tend to break the workflows.
In an attempt to fix that, instead of relying on stable and nightly, we can
instead make it explicit which versions of Rust we want to use.

The proposed solution here is:
- Use 1.82 for formatting, Clippy and for running cargo make binaries.
- Use 1.78 as the MSRV and for running tests.
- nightly-2024-10-18 for cargo-public-api (as per recommended [on docs](https://github.com/cargo-public-api/cargo-public-api))

Why 1.82.0? Just because latest - 2 (latest 1.82). But we can choose another one
if we prefer.